Parent Offset Clipping Image Issue
When hovering the image to produce the zoom effect, the offset of the parent caused by the border on your demo has an additional 5 pixels being cut off of the left part of the image. By removing ".parent()" you can change the location of the images as well as modify borders or margins within the parent container without altering the positioning of the zoom. I modified the variable names to be consistent with the project format.